-Updates to make Juggler compile better under Linux -Rewrote the reconfiguration system -Eliminated several compiler warnings
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@1504 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Input/vjSim/vjSimDigital.h
@@ -46,7 +46,7 @@ public:
    //: Return analog data
    virtual int getDigitalData(int devNum=0)
    {
-      vjASSERT(devNum < mDigitalData.size());    // Make sure we have enough space
+      vjASSERT(devNum < (int)mDigitalData.size());    // Make sure we have enough space
       return mDigitalData[devNum];
    }
